pyproject.toml
[project]
name = "api-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
